Bake OG: worker_threads pool for agency cards — another 3-4× on top
resvg-js's render() is sync and doesn't release the JS thread (verified
by microbench), so Promise.all of N renders doesn't actually go
parallel — only the satori half does. Real multi-core parallelism
needs separate threads, so agency renders now go through a
node:worker_threads pool.

Architecture:
- bake-og-render.mjs: createAgencyBuilder factory, used by both main
  thread (for home + 287g) and each worker (for one agency at a time).
- bake-og-worker.mjs: thin worker entry that initializes the builder
  from workerData (fonts, brand PNG, pre-rasterized base maps,
  jurisdiction/centroid maps) and processes one agency per message.
- bake-og-images.mjs (main): loads shared state once, pre-rasterizes
  the base map per output size, spawns the pool, dispatches.

Pool size defaults to os.cpus().length; cap with BAKE_OG_CONCURRENCY=N
if memory is tight (each worker copies ~3 MB of assets via
structuredClone).

Measured on 16-core box, 100-card sample:
- baseline (pre-sharp): 5.2 sec/card, 8:39 wall
- sharp composite (single-thread): 0.59 sec/card, 1:01 wall
- this change: 0.18 sec/card, 0:21 wall (1047% cpu, ~10.5 cores active)

Full 770-card bake projects to ~2.5 min vs the original 60+ min.
