+++ packages/web/scripts/bake-og-images.mjs
@@ -7,22 +7,30 @@
 //   pnpm bake:og --only home
 //   pnpm bake:og --only agencies
 //   pnpm bake:og --slug missouri-state-highway-patrol
+//   pnpm bake:og --limit 100     # for benchmarking — bake only the first N
+//
+// Concurrency: agency cards are rendered in a node:worker_threads pool.
+// Pool size defaults to os.cpus().length; override with
+// BAKE_OG_CONCURRENCY=N to tune.
 //
 // Output: packages/web/static/og/{home,287g,agency/{slug}}.png (gitignored)
 
 import { createRequire } from "node:module";
-import { mkdir, readFile, writeFile } from "node:fs/promises";
+import { mkdir, readFile } from "node:fs/promises";
 import { fileURLToPath } from "node:url";
+import os from "node:os";
 import path from "node:path";
-import satori from "satori";
-import { html as htmlToVnode } from "satori-html";
+import { Worker } from "node:worker_threads";
 import { Resvg } from "@resvg/resvg-js";
 import sharp from "sharp";
+
+import { createAgencyBuilder } from "./bake-og-render.mjs";
 
 const require = createRequire(import.meta.url);
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const STATIC_DIR = path.resolve(__dirname, "..", "static");
 const OUT_DIR = path.join(STATIC_DIR, "og");
+const WORKER_PATH = path.join(__dirname, "bake-og-worker.mjs");
 
 const CDN_BASE = process.env.PUBLIC_DATA_BASE_URL ?? "https://data.vsr.recoveredfactory.net";
 const RELEASE_PATH = process.env.PUBLIC_DATA_RELEASE_PATH ?? "/releases/v2.1";
@@ -63,15 +71,10 @@ const fonts = [
 
 // ---------- background brand image ----------
 //
-// Held as a raw PNG buffer (not a data URL). renderToPng composites the
-// satori-rendered card over this with sharp, instead of embedding it in
-// the SVG and letting resvg decode it per card — that decode was a
-// dominant cost in the per-card render path.
-//
-// Resized to the card output size on startup so the composite step
-// produces 1200×630 output without sharp having to resize per call.
-// The original CSS used `background-size: 1200px 630px`, so this
-// matches that behavior.
+// Resized once on startup to the card output size. The original CSS
+// used `background-size: 1200px 630px`, so `fit: "fill"` matches that
+// stretch behavior exactly. Held as a PNG buffer (not a data URL) so
+// sharp can use it directly as the bottom of the composite.
 
 const brandPngSrc = await readFile(path.join(STATIC_DIR, "social-meta.png"));
 const brandPng = await sharp(brandPngSrc)
@@ -79,7 +82,7 @@ const brandPng = await sharp(brandPngSrc)
   .png()
   .toBuffer();
 
-// ---------- map svg: parse once, render per-agency on demand ----------
+// ---------- map svg: parse once ----------
 
 const fetchText = async (url) => {
   const r = await fetch(url);
@@ -89,8 +92,6 @@ const fetchText = async (url) => {
 
 const locatorSvg = await fetchText(`${CDN_BASE}${RELEASE_PATH}/dist/mo_locator.svg`);
 
-// Pull state outline, the roads group, and per-agency jurisdiction
-// paths from the source SVG so we can render any subset we like.
 const viewBoxMatch = locatorSvg.match(/viewBox="([^"]+)"/);
 const statePathMatch = locatorSvg.match(/<path class="state" d="([^"]+)"/);
 if (!viewBoxMatch || !statePathMatch) {
@@ -137,9 +138,6 @@ let jm;
 while ((jm = jurisRe.exec(locatorSvg))) {
   const slug = jm[1];
   const d = jm[2];
-  // Pull every numeric coordinate out of the path. The locator SVG only
-  // uses absolute M/L commands with space-separated x y pairs, so we can
-  // just collect all signed decimals and pair them up.
   const nums = d.match(/-?\d+(?:\.\d+)?/g);
   let w = 0;
   let h = 0;
@@ -159,7 +157,6 @@ while ((jm = jurisRe.exec(locatorSvg))) {
   jurisdictionBySlug.set(slug, { d, maxDim: Math.max(w, h) });
 }
 
-// Centroid points by slug.
 const centroidByslug = new Map();
 const circleRe = /<circle[^>]*data-slug="([^"]+)"[^>]*\bcx="([^"]+)"[^>]*\bcy="([^"]+)"/g;
 let m;
@@ -167,283 +164,61 @@ while ((m = circleRe.exec(locatorSvg))) {
   centroidByslug.set(m[1], { cx: Number(m[2]), cy: Number(m[3]) });
 }
 
-// Pre-rasterize the base map (state outline + roads) once per output
-// size as a raw RGBA pixel buffer. Per-agency renders only have to
-// rasterize the tiny highlight (one circle or one polygon) and sharp
-// composites that onto a copy of the cached base. Measured ~1.5×
-// faster than re-rasterizing the full source per call.
+// ---------- pre-rasterize the base map per size ----------
 //
-// Why not let resvg do the composite via <image href="..."> on a data
-// URL of the cached base? Tried it — was ~6× slower than no caching
-// at all. resvg's PNG decode + base64 path costs more than parsing the
-// source paths from scratch. The composite has to happen at the pixel
-// level, outside of SVG.
+// Used sizes: 300 (home + per-agency inset) and 260 (287g inset).
+// Each base is held as both a raw RGBA buffer (for sharp composite when
+// a highlight needs to go on top) and a finished PNG data URL (for the
+// "no highlight" path).
+
 const BASE_BG = "#ffffff";
-const baseRasterBySize = new Map();
-const getBaseRaster = (pxSize) => {
-  const cached = baseRasterBySize.get(pxSize);
-  if (cached) return cached;
-  const svg = `<?xml version="1.0" encoding="UTF-8"?>
+const USED_SIZES = [260, 300];
+
+const buildBaseSvg = () => `<?xml version="1.0" encoding="UTF-8"?>
 <svg xmlns="http://www.w3.org/2000/svg" viewBox="${VIEWBOX}" preserveAspectRatio="xMidYMid meet">
   <path d="${STATE_PATH}" fill="#f1f5f9" stroke="#0f172a" stroke-width="0.025" stroke-linejoin="round" />
   ${ROADS_SVG}
 </svg>`;
-  const img = new Resvg(svg, {
+
+const baseRasters = {};
+const baseDataUrls = {};
+for (const pxSize of USED_SIZES) {
+  const img = new Resvg(buildBaseSvg(), {
     fitTo: { mode: "width", value: pxSize },
     background: BASE_BG,
   }).render();
-  const entry = {
-    pixels: Buffer.from(img.pixels),
-    width: img.width,
-    height: img.height,
-  };
-  baseRasterBySize.set(pxSize, entry);
-  return entry;
-};
-
-const baseAsPngBufferBySize = new Map();
-const getBaseAsPngBuffer = async (pxSize) => {
-  const cached = baseAsPngBufferBySize.get(pxSize);
-  if (cached) return cached;
-  const base = getBaseRaster(pxSize);
-  const buf = await sharp(base.pixels, {
-    raw: { width: base.width, height: base.height, channels: 4 },
+  const pixels = Buffer.from(img.pixels);
+  baseRasters[pxSize] = { pixels, width: img.width, height: img.height };
+  const pngBuf = await sharp(pixels, {
+    raw: { width: img.width, height: img.height, channels: 4 },
   })
     .png()
     .toBuffer();
-  baseAsPngBufferBySize.set(pxSize, buf);
-  return buf;
+  baseDataUrls[pxSize] = `data:image/png;base64,${pngBuf.toString("base64")}`;
+}
+
+// ---------- shared render env ----------
+//
+// One bundle of state passed both into the in-process builder (used for
+// home + 287g) and into each worker thread (used for agency cards).
+// Buffers and Maps clone cleanly through structuredClone for workerData.
+
+const renderEnv = {
+  fonts,
+  brandPng,
+  viewbox: VIEWBOX,
+  baseRasters,
+  baseDataUrls,
+  jurisdictionEntries: [...jurisdictionBySlug.entries()],
+  centroidEntries: [...centroidByslug.entries()],
+  outDir: OUT_DIR,
+  size: SIZE,
 };
 
-// Render a mini-map at the given pixel size. `mode`:
-//   "none"        — state outline + roads, no highlight
-//   "dots"        — state + roads + a list of centroid dots
-//   "jurisdiction"— state + roads + one filled polygon
-const renderMapPng = async (opts, pxSize = 300) => {
-  const { mode = "none", dots = [], path = null } = opts ?? {};
+const mainBuilder = createAgencyBuilder(renderEnv);
+const { renderMapPng, renderToPng, card, writePng } = mainBuilder;
 
-  let highlight = "";
-  if (mode === "jurisdiction" && path) {
-    highlight = `<path d="${path}" fill="#047857" stroke="#065f46" stroke-width="0.025" stroke-linejoin="round" opacity="0.9" />`;
-  } else if (mode === "dots" && dots.length) {
-    const n = dots.length;
-    const r = n <= 1 ? 0.18 : n <= 20 ? 0.13 : n <= 80 ? 0.1 : 0.08;
-    const stroke = n <= 1 ? 0.05 : 0.025;
-    highlight = dots
-      .map(
-        ({ cx, cy }) =>
-          `<circle cx="${cx}" cy="${cy}" r="${r}" fill="#047857" stroke="#ffffff" stroke-width="${stroke}" />`,
-      )
-      .join("");
-  }
-  if (!highlight) {
-    const buf = await getBaseAsPngBuffer(pxSize);
-    return `data:image/png;base64,${buf.toString("base64")}`;
-  }
-
-  const base = getBaseRaster(pxSize);
-  // Highlight-only SVG with the same viewBox so the rasterized output
-  // aligns pixel-for-pixel with the base. Transparent background means
-  // only the drawn shape contributes alpha.
-  const overlaySvg = `<?xml version="1.0" encoding="UTF-8"?>
-<svg xmlns="http://www.w3.org/2000/svg" viewBox="${VIEWBOX}" preserveAspectRatio="xMidYMid meet">
-  ${highlight}
-</svg>`;
-  const overlay = new Resvg(overlaySvg, {
-    fitTo: { mode: "width", value: pxSize },
-  }).render();
-
-  const composedPng = await sharp(base.pixels, {
-    raw: { width: base.width, height: base.height, channels: 4 },
-  })
-    .composite([
-      {
-        input: Buffer.from(overlay.pixels),
-        raw: { width: overlay.width, height: overlay.height, channels: 4 },
-        blend: "over",
-      },
-    ])
-    .png()
-    .toBuffer();
-  return `data:image/png;base64,${composedPng.toString("base64")}`;
-};
-
-// Pre-render the no-highlight base map once — used for the home card and
-// as a fallback when an agency has neither a jurisdiction nor a centroid.
-const baseMapDataUrl = await renderMapPng({ mode: "none" });
-
-// ---------- helpers ----------
-
-const formatStops = (n) => {
-  if (!Number.isFinite(n)) return "—";
-  return new Intl.NumberFormat("en-US").format(Math.round(n));
-};
-
-// HTML-parser safety only. Apostrophes and double quotes are valid
-// inside text nodes, so leaving them alone keeps "Sheriff's Office"
-// rendering correctly instead of as "Sheriff&#39;s Office".
-const escapeHtml = (s) =>
-  String(s).replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
-
-// Pick a title font size based on character count so long agency names
-// don't overflow the (~660 px) text column. Tuned so most titles wrap
-// to two lines and the stat block always fits below.
-const titleSize = (s) => {
-  const n = s.length;
-  if (n <= 16) return 82;
-  if (n <= 22) return 66;
-  if (n <= 30) return 56;
-  if (n <= 40) return 48;
-  if (n <= 50) return 42;
-  return 36;
-};
-
-// ---------- card template ----------
-
-const card = ({
-  eyebrow,
-  title,
-  subtitle,
-  stat,
-  statLabel,
-  mapDataUrl,
-}) => `
-<div style="
-  display: flex;
-  width: 1200px;
-  height: 630px;
-  padding: 65px 105px;
-  font-family: 'Inter';
-">
-  <div style="
-    display: flex;
-    flex-direction: column;
-    width: 990px;
-    height: 500px;
-    background: #ffffff;
-    border-radius: 28px;
-    padding: 52px 56px;
-    box-shadow: 0 24px 48px rgba(15, 23, 42, 0.25);
-  ">
-    <div style="
-      display: flex;
-      font-size: 32px;
-      font-weight: 700;
-      letter-spacing: 0.18em;
-      color: #047857;
-      text-transform: uppercase;
-    ">
-      ${escapeHtml(eyebrow)}
-    </div>
-
-    <div style="display: flex; flex-grow: 1; padding-top: 32px;">
-      <div style="
-        display: flex;
-        flex-direction: column;
-        flex: 1 1 0;
-        min-width: 0;
-        padding-right: 40px;
-      ">
-        <div style="
-          display: flex;
-          font-size: ${titleSize(title)}px;
-          font-weight: 900;
-          line-height: 1.05;
-          color: #0f172a;
-          letter-spacing: -0.02em;
-        ">
-          ${escapeHtml(title)}
-        </div>
-        ${
-          subtitle
-            ? `<div style="
-                display: flex;
-                margin-top: 14px;
-                font-size: 28px;
-                font-weight: 500;
-                color: #475569;
-              ">${escapeHtml(subtitle)}</div>`
-            : ""
-        }
-
-        <div style="display: flex; flex-grow: 1;"></div>
-
-        ${
-          stat
-            ? `<div style="display: flex; flex-direction: column;">
-                <div style="
-                  display: flex;
-                  font-size: 116px;
-                  font-weight: 900;
-                  line-height: 1;
-                  color: #047857;
-                  letter-spacing: -0.03em;
-                ">${escapeHtml(stat)}</div>
-                ${
-                  statLabel
-                    ? `<div style="display: flex; margin-top: 8px; font-size: 26px; color: #475569;">${escapeHtml(statLabel)}</div>`
-                    : ""
-                }
-              </div>`
-            : ""
-        }
-      </div>
-
-      <div style="
-        display: flex;
-        flex-direction: column;
-        width: 280px;
-        flex-shrink: 0;
-        align-items: flex-end;
-        justify-content: space-between;
-        padding-top: 4px;
-      ">
-        ${
-          mapDataUrl
-            ? `<img src="${mapDataUrl}" style="width: 270px; height: 270px;" />`
-            : `<div style="display: flex; width: 270px; height: 270px;"></div>`
-        }
-        <div style="display: flex; font-size: 20px; font-weight: 600; color: #94a3b8;">
-          vsr.recoveredfactory.net
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-`;
-
-// ---------- render pipeline ----------
-
-const renderToPng = async (htmlString) => {
-  const vnode = htmlToVnode(htmlString);
-  const svg = await satori(vnode, { ...SIZE, fonts });
-  // Render the card to a raw RGBA buffer with a transparent background
-  // (the outer div has no fill, so only the white panel + box-shadow
-  // contribute pixels). Then sharp composites that over the brand PNG.
-  // This skips embedding the 1.2 MB brand data URL in the satori SVG —
-  // resvg used to re-parse + decode it on every card.
-  const cardImg = new Resvg(svg, {
-    fitTo: { mode: "width", value: SIZE.width },
-  }).render();
-  return sharp(brandPng)
-    .composite([
-      {
-        input: Buffer.from(cardImg.pixels),
-        raw: { width: cardImg.width, height: cardImg.height, channels: 4 },
-        blend: "over",
-      },
-    ])
-    .png()
-    .toBuffer();
-};
-
-const writePng = async (relPath, buffer) => {
-  const out = path.join(OUT_DIR, relPath);
-  await mkdir(path.dirname(out), { recursive: true });
-  await writeFile(out, buffer);
-};
-
-// ---------- per-page builders ----------
+// ---------- per-page builders (main thread) ----------
 
 const buildHome = async () => {
   // Use the real agency count for the stat — the map only shows agencies
@@ -508,7 +283,7 @@ const build287g = async () => {
 
   const mapDataUrl = highlights.length
     ? await renderMapPng({ mode: "dots", dots: highlights }, 260)
-    : baseMapDataUrl;
+    : baseDataUrls[300];
 
   const png = await renderToPng(
     card({
@@ -526,62 +301,7 @@ const build287g = async () => {
   );
 };
 
-const buildAgency = async (agency) => {
-  const name = agency.canonical_name || agency.agency_slug;
-  const isStateAgency = agency.agency_type === "State Agency";
-
-  // Subtitle: county only — town gets dropped per request. Skip
-  // entirely for state agencies (their reach is the whole map).
-  const subtitle = isStateAgency ? null : agency.county || null;
-
-  const stops = Number(agency.stops ?? agency.all_stops_total ?? 0);
-  const latestYear =
-    Array.isArray(agency.years_with_data) && agency.years_with_data.length
-      ? Math.max(...agency.years_with_data.map(Number).filter(Number.isFinite))
-      : null;
-
-  // Map highlight priority: jurisdiction polygon (county/city outline)
-  // → centroid dot → nothing. State agencies always get the bare map.
-  // Tiny polygons (small-town city limits) fall through to the centroid
-  // because their fill would render sub-pixel in a 300 px inset.
-  // Threshold is in viewBox units; the locator viewBox is ~5.24 wide, so
-  // 0.1 units ≈ 6 px at 300 px output — below that the polygon is
-  // invisible and the dot is what readers actually see.
-  const MIN_VISIBLE_DIM = 0.1;
-  let mapDataUrl;
-  if (isStateAgency) {
-    mapDataUrl = baseMapDataUrl;
-  } else {
-    const jurisdiction = jurisdictionBySlug.get(agency.agency_slug);
-    const centroid = centroidByslug.get(agency.agency_slug);
-    const jurisdictionVisible = jurisdiction && jurisdiction.maxDim >= MIN_VISIBLE_DIM;
-    if (jurisdictionVisible) {
-      mapDataUrl = await renderMapPng({ mode: "jurisdiction", path: jurisdiction.d }, 300);
-    } else if (centroid) {
-      mapDataUrl = await renderMapPng({ mode: "dots", dots: [centroid] }, 300);
-    } else if (jurisdiction) {
-      // Polygon exists but is sub-pixel and there's no centroid — render
-      // it anyway rather than dropping the highlight entirely.
-      mapDataUrl = await renderMapPng({ mode: "jurisdiction", path: jurisdiction.d }, 300);
-    } else {
-      mapDataUrl = baseMapDataUrl;
-    }
-  }
-
-  const png = await renderToPng(
-    card({
-      eyebrow: "Missouri Vehicle Stops",
-      title: name,
-      subtitle,
-      stat: stops > 0 ? formatStops(stops) : null,
-      statLabel: stops > 0
-        ? `stops${latestYear ? ` in ${latestYear}` : ""}`
-        : null,
-      mapDataUrl,
-    }),
-  );
-  await writePng(`agency/${agency.agency_slug}.png`, png);
-};
+// ---------- agency cards via worker pool ----------
 
 const buildAllAgencies = async () => {
   const url = `${CDN_BASE}${RELEASE_PATH}/dist/agency_index.json`;
@@ -598,28 +318,67 @@ const buildAllAgencies = async () => {
   if (limit !== null) {
     agencies = agencies.slice(0, limit);
   }
-  console.log(`Baking ${agencies.length} agency cards…`);
+
+  // Worker pool: each worker pays the per-process startup (font load,
+  // sharp init) once and then handles many cards. resvg-js's render()
+  // blocks the JS thread, so true parallelism only comes from running
+  // multiple Node threads — hence worker_threads, not Promise.all.
+  const envOverride = Number(process.env.BAKE_OG_CONCURRENCY);
+  const poolSize = Number.isFinite(envOverride) && envOverride > 0
+    ? Math.min(Math.floor(envOverride), agencies.length)
+    : Math.min(os.cpus().length, agencies.length);
+
+  console.log(
+    `Baking ${agencies.length} agency cards across ${poolSize} worker${poolSize === 1 ? "" : "s"}…`,
+  );
   const t0 = Date.now();
+
+  const workers = Array.from(
+    { length: poolSize },
+    () => new Worker(WORKER_PATH, { workerData: renderEnv }),
+  );
+
   let done = 0;
-  // resvg-js's render() is synchronous and blocks the JS thread —
-  // verified by microbench (Promise.all of N renders ≈ sequential N).
-  // Increasing this past a small number buys nothing and adds memory
-  // pressure; the only thing it interleaves is satori's async layout
-  // step. For real parallelism we'd need worker_threads.
-  const CONCURRENCY = 6;
   const queue = [...agencies];
-  await Promise.all(
-    Array.from({ length: CONCURRENCY }, async () => {
-      while (queue.length) {
-        const a = queue.shift();
-        await buildAgency(a);
+
+  const runWorker = (w) =>
+    new Promise((resolveWorker, rejectWorker) => {
+      const dispatchNext = () => {
+        if (!queue.length) {
+          w.postMessage({ shutdown: true });
+          return;
+        }
+        const agency = queue.shift();
+        w.postMessage({ agency });
+      };
+
+      w.on("message", (msg) => {
+        if (msg?.error) {
+          rejectWorker(new Error(msg.error));
+          return;
+        }
         done += 1;
         if (done % 25 === 0 || done === agencies.length) {
           process.stdout.write(`  ${done}/${agencies.length}\r`);
         }
-      }
-    }),
-  );
+        dispatchNext();
+      });
+      w.on("error", rejectWorker);
+      w.on("exit", (code) => {
+        if (code === 0) resolveWorker();
+        else rejectWorker(new Error(`worker exited with code ${code}`));
+      });
+
+      dispatchNext();
+    });
+
+  try {
+    await Promise.all(workers.map(runWorker));
+  } catch (err) {
+    for (const w of workers) w.terminate();
+    throw err;
+  }
+
   const elapsed = ((Date.now() - t0) / 1000).toFixed(1);
   console.log(`\n✓ ${agencies.length} agency cards in ${elapsed}s`);
 };
